Refactor: Change types in DTO's

+++ src/main/java/com/example/Product_Management/dto/ProductRequestDTO.java
@@ -1,9 +1,7 @@
 package com.example.Product_Management.dto;
 
-import java.math.BigDecimal;
-
 public record ProductRequestDTO(
         String nameProduct,
         String descriptionProduct,
-        BigDecimal priceProduct
+        Double priceProduct
 ) {}
