+++ src/main/java/com/example/Product_Management/dto/ProductResponseDTO.java
@@ -1,11 +1,10 @@
 package com.example.Product_Management.dto;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record ProductResponseDTO(
         UUID id,
         String name,
         String description,
-        BigDecimal priceProduct
+        Double priceProduct
 ) {}
